ajout titre et modif titre

+++ src/Controller/UserController.php
@@ -47,8 +47,8 @@ class UserController extends AbstractController
     public function show(User $user, SpacesRepository $spacesRepository): Response
     {
 
-        if ($user !== $this->getUser() && !$this->isGranted('ROLE_USER')) {
-            throw $this->createAccessDeniedException('Vous n\'êtes pas autorisé à modifier ce profil.');
+        if ($user !== $this->getUser() && !$this->isGranted('ROLE_USER') && !$this->isGranted('ROLE_ADMIN')) {
+            throw $this->createAccessDeniedException('Vous n\'êtes pas autorisé à voir ce profil.');
         }
 
         $connectedSpace = $spacesRepository->findConnectedSpaceForUser($user);
@@ -62,12 +62,15 @@ class UserController extends AbstractController
     #[Route('/{id}/edit', name: 'app_user_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, User $user, EntityManagerInterface $entityManager): Response
     {
+        if ($user !== $this->getUser() && !$this->isGranted('ROLE_USER') && !$this->isGranted('ROLE_ADMIN')) {
+            throw $this->createAccessDeniedException('Vous n\'êtes pas autorisé à modifier ce profil.');
+        }
+
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-//            $user->setRoles($form->get('roles')->getData());
-//            $entityManager->persist($user);
+            $entityManager->persist($user);
             $entityManager->flush();
 
             return $this->redirectToRoute('app_user_index', [], Response::HTTP_SEE_OTHER);
@@ -82,6 +85,10 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'app_user_delete', methods: ['POST'])]
     public function delete(Request $request, User $user, EntityManagerInterface $entityManager): Response
     {
+        if ($user !== $this->getUser() && !$this->isGranted('ROLE_USER') && !$this->isGranted('ROLE_ADMIN')) {
+            throw $this->createAccessDeniedException('Vous n\'êtes pas autorisé à modifier ce profil.');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
             $entityManager->remove($user);
             $entityManager->flush();
